requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.11
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.1
django-multiselectfield==0.1.12
django-rest-swagger==2.2.0
django-utils-six==2.0
djangorestframework==3.14.0
future==1.0.0
gunicorn==22.0.0
idna==3.6
importlib_metadata==7.0.2
iso8601==2.1.0
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
mpyg321==2.2.1
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==24.0
panda==0.3.1
pexpect==4.9.0
playsound==1.3.0
psutil==5.9.8
ptyprocess==0.7.0
pyserial==3.5
python-memcached==1.62
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
singleton-decorator==1.0.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.6
whitenoise==6.6.0
zipp==3.17.0
